requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
pymysql==1.1.1
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.43
requests==2.32.3